move type definition to common
