+++ client/query/get_depth.go
@@ -4,42 +4,11 @@ import (
 	"encoding/json"
 
 	"github.com/binance-chain/go-sdk/common"
+	"github.com/binance-chain/go-sdk/common/types"
 )
 
-// DepthQuery
-type DepthQuery struct {
-	Symbol string  `json:"symbol"`
-	Limit  *uint32 `json:"limit,omitempty,string"`
-}
-
-func NewDepthQuery(baseAssetSymbol, quoteAssetSymbol string) *DepthQuery {
-	return &DepthQuery{Symbol: common.CombineSymbol(baseAssetSymbol, quoteAssetSymbol)}
-}
-
-func (param *DepthQuery) WithLimit(limit uint32) *DepthQuery {
-	param.Limit = &limit
-	return param
-}
-
-func (param *DepthQuery) Check() error {
-	if param.Symbol == "" {
-		return SymbolMissingError
-	}
-	if param.Limit != nil && *param.Limit <= 0 {
-		return LimitOutOfRangeError
-	}
-	return nil
-}
-
-// MarketDepth broad caste to the user
-type MarketDepth struct {
-	Bids   [][]string `json:"bids"` // "bids": [ [ "0.0024", "10" ] ]
-	Asks   [][]string `json:"asks"` // "asks": [ [ "0.0024", "10" ] ]
-	Height int64      `json:"height"`
-}
-
 // GetDepth returns market depth records
-func (c *client) GetDepth(query *DepthQuery) (*MarketDepth, error) {
+func (c *client) GetDepth(query *types.DepthQuery) (*types.MarketDepth, error) {
 	err := query.Check()
 	if err != nil {
 		return nil, err
@@ -53,7 +22,7 @@ func (c *client) GetDepth(query *DepthQuery) (*MarketDepth, error) {
 		return nil, err
 	}
 
-	var MarketDepth MarketDepth
+	var MarketDepth types.MarketDepth
 	if err := json.Unmarshal(resp, &MarketDepth); err != nil {
 		return nil, err
 	}
