+++ client/query/query.go
@@ -6,18 +6,18 @@ import (
 )
 
 type QueryClient interface {
-	GetClosedOrders(query *ClosedOrdersQuery) (*CloseOrders, error)
-	GetDepth(query *DepthQuery) (*MarketDepth, error)
-	GetKlines(query *KlineQuery) ([]Kline, error)
-	GetMarkets(query *MarketsQuery) ([]types.TradingPair, error)
-	GetOrder(orderID string) (*Order, error)
-	GetOpenOrders(query *OpenOrdersQuery) (*OpenOrders, error)
-	GetTicker24h(query *Ticker24hQuery) ([]Ticker24h, error)
-	GetTrades(query *TradesQuery) (*Trades, error)
-	GetAccount(string) (*Account, error)
-	GetTime() (*Time, error)
+	GetClosedOrders(query *types.ClosedOrdersQuery) (*types.CloseOrders, error)
+	GetDepth(query *types.DepthQuery) (*types.MarketDepth, error)
+	GetKlines(query *types.KlineQuery) ([]types.Kline, error)
+	GetMarkets(query *types.MarketsQuery) ([]types.TradingPair, error)
+	GetOrder(orderID string) (*types.Order, error)
+	GetOpenOrders(query *types.OpenOrdersQuery) (*types.OpenOrders, error)
+	GetTicker24h(query *types.Ticker24hQuery) ([]types.Ticker24h, error)
+	GetTrades(query *types.TradesQuery) (*types.Trades, error)
+	GetAccount(string) (*types.BalanceAccount, error)
+	GetTime() (*types.Time, error)
 	GetTokens() ([]types.Token, error)
-	GetNodeInfo() (*ResultStatus, error)
+	GetNodeInfo() (*types.ResultStatus, error)
 }
 
 type client struct {
