+++ client/rpc/validate.go
@@ -3,29 +3,37 @@ package rpc
 import (
 	"crypto/sha256"
 	"fmt"
-	"github.com/pkg/errors"
 	"github.com/tendermint/tendermint/libs/common"
 	"github.com/tendermint/tendermint/types"
+	"strings"
 )
 
 const (
 	maxABCIPathLength     = 1024
 	maxABCIDataLength     = 1024 * 1024
 	maxTxLength           = 1024 * 1024
-	maxCommonStringLength = 1024
+	maxABCIQueryStrLength = 1024
+	maxTxSearchStrLength  = 1024
 	maxUnConfirmedTxs     = 100
+
+	tokenSymbolMaxLen = 11
+	tokenSymbolMinLen = 3
 )
 
 var (
-	ExceedABCIPathLengthError       = fmt.Errorf("the abci path exceed max length %d ", maxABCIPathLength)
-	ExceedABCIDataLengthError       = fmt.Errorf("the abci data exceed max length %d ", maxABCIDataLength)
-	ExceedTxLengthError             = fmt.Errorf("the tx data exceed max length %d ", maxTxLength)
-	LimitNegativeError              = fmt.Errorf("the limit can't be negative")
-	ExceedMaxUnConfirmedTxsNumError = fmt.Errorf("the limit of unConfirmed tx exceed max limit %d ", maxUnConfirmedTxs)
-	HeightNegativeError             = fmt.Errorf("the height can't be negative")
-	MaxMinHeightConflictError       = fmt.Errorf("the min height can't be larger than max height")
-	HashLengthError                 = fmt.Errorf("the length of hash is not 32")
-	ExceedCommonStrLengthError      = fmt.Errorf("the query string exceed max length %d ", maxABCIPathLength)
+	ExceedABCIPathLengthError         = fmt.Errorf("the abci path exceed max length %d ", maxABCIPathLength)
+	ExceedABCIDataLengthError         = fmt.Errorf("the abci data exceed max length %d ", maxABCIDataLength)
+	ExceedTxLengthError               = fmt.Errorf("the tx data exceed max length %d ", maxTxLength)
+	LimitNegativeError                = fmt.Errorf("the limit can't be negative")
+	ExceedMaxUnConfirmedTxsNumError   = fmt.Errorf("the limit of unConfirmed tx exceed max limit %d ", maxUnConfirmedTxs)
+	HeightNegativeError               = fmt.Errorf("the height can't be negative")
+	MaxMinHeightConflictError         = fmt.Errorf("the min height can't be larger than max height")
+	HashLengthError                   = fmt.Errorf("the length of hash is not 32")
+	ExceedABCIQueryStrLengthError     = fmt.Errorf("the query string exceed max length %d ", maxABCIPathLength)
+	ExceedTxSearchQueryStrLengthError = fmt.Errorf("the query string exceed max length %d ", maxTxSearchStrLength)
+	OffsetNegativeError               = fmt.Errorf("offset can't be less than 0")
+	SymbolLengthExceedRangeError      = fmt.Errorf("length of symbol should be in range [%d,%d]", tokenSymbolMinLen, tokenSymbolMaxLen)
+	PairFormatError                   = fmt.Errorf("the pair should in format 'symbol1_symbol2'")
 )
 
 func ValidateABCIPath(path string) error {
@@ -62,7 +70,7 @@ func ValidateHeightRange(from int64, to int64) error {
 	if from < 0 || to < 0 {
 		return HeightNegativeError
 	}
-	if from < to {
+	if from > to {
 		return MaxMinHeightConflictError
 	}
 	return nil
@@ -82,9 +90,51 @@ func ValidateHash(hash []byte) error {
 	return nil
 }
 
-func ValidateCommonStr(query string) error {
-	if len(query) > maxCommonStringLength {
-		return ExceedCommonStrLengthError
+func ValidateABCIQueryStr(query string) error {
+	if len(query) > maxABCIQueryStrLength {
+		return ExceedABCIQueryStrLengthError
+	}
+	return nil
+}
+
+func ValidateTxSearchQueryStr(query string) error {
+	if len(query) > maxTxSearchStrLength {
+		return ExceedTxSearchQueryStrLengthError
+	}
+	return nil
+}
+
+func ValidateOffset(offset int) error {
+	if offset < 0 {
+		return OffsetNegativeError
+	}
+	return nil
+}
+
+func ValidateLimit(limit int) error {
+	if limit < 0 {
+		return LimitNegativeError
+	}
+	return nil
+}
+
+func ValidateSymbol(symbol string) error {
+	if len(symbol) > tokenSymbolMaxLen || len(symbol) < tokenSymbolMinLen {
+		return SymbolLengthExceedRangeError
+	}
+	return nil
+}
+
+func ValidatePair(pair string) error {
+	symbols := strings.Split(pair, "_")
+	if len(symbols) != 2 {
+		return PairFormatError
+	}
+	if err := ValidateSymbol(symbols[0]); err != nil {
+		return err
+	}
+	if err := ValidateSymbol(symbols[1]); err != nil {
+		return err
 	}
 	return nil
 }
