+++ common/types/account.go
@@ -1,21 +1,9 @@
 package types
 
 import (
-	"strings"
-
 	"github.com/pkg/errors"
 	"github.com/tendermint/tendermint/crypto"
 )
-
-// Token definition
-type Token struct {
-	Name        string     `json:"name"`
-	Symbol      string     `json:"symbol"`
-	OrigSymbol  string     `json:"original_symbol"`
-	TotalSupply Fixed8     `json:"total_supply"`
-	Owner       AccAddress `json:"owner"`
-	Mintable    bool       `json:"mintable"`
-}
 
 // AppAccount definition
 type AppAccount struct {
@@ -23,153 +11,6 @@ type AppAccount struct {
 	Name        string `json:"name"`
 	FrozenCoins Coins  `json:"frozen"`
 	LockedCoins Coins  `json:"locked"`
-}
-
-// Coin def
-// Coin def
-type Coin struct {
-	Denom  string `json:"denom"`
-	Amount int64  `json:"amount"`
-}
-
-func (coin Coin) IsZero() bool {
-	return coin.Amount == 0
-}
-
-func (coin Coin) IsPositive() bool {
-	return coin.Amount > 0
-}
-
-func (coin Coin) IsNotNegative() bool {
-	return coin.Amount >= 0
-}
-
-func (coin Coin) SameDenomAs(other Coin) bool {
-	return (coin.Denom == other.Denom)
-}
-
-func (coin Coin) Plus(coinB Coin) Coin {
-	if !coin.SameDenomAs(coinB) {
-		return coin
-	}
-	return Coin{coin.Denom, coin.Amount + coinB.Amount}
-}
-
-// Coins def
-type Coins []Coin
-
-func (coins Coins) IsValid() bool {
-	switch len(coins) {
-	case 0:
-		return true
-	case 1:
-		return !coins[0].IsZero()
-	default:
-		lowDenom := coins[0].Denom
-		for _, coin := range coins[1:] {
-			if coin.Denom <= lowDenom {
-				return false
-			}
-			if coin.IsZero() {
-				return false
-			}
-			lowDenom = coin.Denom
-		}
-		return true
-	}
-}
-
-func (coins Coins) IsPositive() bool {
-	if len(coins) == 0 {
-		return false
-	}
-	for _, coin := range coins {
-		if !coin.IsPositive() {
-			return false
-		}
-	}
-	return true
-}
-
-func (coins Coins) Plus(coinsB Coins) Coins {
-	sum := ([]Coin)(nil)
-	indexA, indexB := 0, 0
-	lenA, lenB := len(coins), len(coinsB)
-	for {
-		if indexA == lenA {
-			if indexB == lenB {
-				return sum
-			}
-			return append(sum, coinsB[indexB:]...)
-		} else if indexB == lenB {
-			return append(sum, coins[indexA:]...)
-		}
-		coinA, coinB := coins[indexA], coinsB[indexB]
-		switch strings.Compare(coinA.Denom, coinB.Denom) {
-		case -1:
-			sum = append(sum, coinA)
-			indexA++
-		case 0:
-			if coinA.Amount+coinB.Amount == 0 {
-				// ignore 0 sum coin type
-			} else {
-				sum = append(sum, coinA.Plus(coinB))
-			}
-			indexA++
-			indexB++
-		case 1:
-			sum = append(sum, coinB)
-			indexB++
-		}
-	}
-}
-
-// IsEqual returns true if the two sets of Coins have the same value
-func (coins Coins) IsEqual(coinsB Coins) bool {
-	if len(coins) != len(coinsB) {
-		return false
-	}
-	for i := 0; i < len(coins); i++ {
-		if coins[i].Denom != coinsB[i].Denom || !(coins[i].Amount == coinsB[i].Amount) {
-			return false
-		}
-	}
-	return true
-}
-
-func (coins Coins) IsNotNegative() bool {
-	if len(coins) == 0 {
-		return true
-	}
-	for _, coin := range coins {
-		if !coin.IsNotNegative() {
-			return false
-		}
-	}
-	return true
-}
-
-func (coins Coins) AmountOf(denom string) int64 {
-	switch len(coins) {
-	case 0:
-		return 0
-	case 1:
-		coin := coins[0]
-		if coin.Denom == denom {
-			return coin.Amount
-		}
-		return 0
-	default:
-		midIdx := len(coins) / 2 // 2:1, 3:1, 4:2
-		coin := coins[midIdx]
-		if denom < coin.Denom {
-			return coins[:midIdx].AmountOf(denom)
-		} else if denom == coin.Denom {
-			return coin.Amount
-		} else {
-			return coins[midIdx+1:].AmountOf(denom)
-		}
-	}
 }
 
 type Account interface {
@@ -341,9 +182,11 @@ func (acc *BaseAccount) Clone() Account {
 	return clonedAcc
 }
 
-type TokenBalance struct {
-	Symbol string `json:"symbol"`
-	Free   Fixed8 `json:"free"`
-	Locked Fixed8 `json:"locked"`
-	Frozen Fixed8 `json:"frozen"`
+// Balance Account definition
+type BalanceAccount struct {
+	Number    int64          `json:"account_number"`
+	Address   string         `json:"address"`
+	Balances  []TokenBalance `json:"balances"`
+	PublicKey []uint8        `json:"public_key"`
+	Sequence  int64          `json:"sequence"`
 }
