+++ common/types/time.go
@@ -1,0 +1,6 @@
+package types
+
+type Time struct {
+	ApTime    string `json:"ap_time"`
+	BlockTime string `json:"block_time"`
+}
